add middleware

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\Http\Middleware\UserAkses;
 use App\Http\Controllers\Admin\SopController;
 use App\Http\Controllers\auth\AuthController;
 use App\Http\Controllers\Admin\TaskController;
@@ -9,6 +10,7 @@ use App\Http\Controllers\Admin\BarcodeController;
 use App\Http\Controllers\Admin\PegawaiController;
 use App\Http\Controllers\Admin\DashboardController;
 use App\Http\Controllers\Admin\AttendancesController;
+use App\Http\Controllers\Users\HomeController;
 use SebastianBergmann\CodeCoverage\Report\Html\Dashboard;
 
 // Route::get('/', function () {
@@ -17,7 +19,7 @@ use SebastianBergmann\CodeCoverage\Report\Html\Dashboard;
 
 Route::get('/test', function () {
     return view('layout.app');
-});
+})->middleware(UserAkses::class);
 
 Route::middleware(['guest'])->group(function () {
     Route::get('/register', [AuthController::class, 'registerView'])->name('register');
@@ -26,57 +28,64 @@ Route::middleware(['guest'])->group(function () {
     Route::get('/login', [AuthController::class, 'loginView'])->name('login');
     Route::post('/login/submit', [AuthController::class, 'loginSubmit'])->name('login.post');
 });  
+Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
+Route::get('/home', [HomeController::class, 'index'])->name('home.pegawai');
 
+Route::middleware('auth','admin:admin')->group(function () {
+    // Route::get('/', function(){
+    //     return redirect('/login');
+    // });
 
-Route::middleware(['auth'])->group(function () {
-    Route::get('/', [DashboardController::class, 'index'])->name('dashboard');
+    
+
+    Route::get('/admin', [DashboardController::class, 'index'])->name('dashboard');
     // Route::post('/login/submit', [AuthController::class, 'loginSubmit'])->name('login.post');
-    Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
+    
 
     Route::controller(PegawaiController::class)->group(function () {
 
         // PEGAWAI ROUTE
-        Route::get('/pegawai','PegawaiIndex')->name('pegawai');
-        Route::get('/pegawai/tambah','PegawaitampilTambah')->name('pegawai.tambah');
-        Route::post('/pegawai/tambah','PegawaiStore')->name('pegawai.post');
-        Route::get('/pegawai/edit/{id}', 'PegawaiEdit')->name('pegawai.edit');
-        Route::put('/pegawai/edit/{id}', 'PegawaiUpdate')->name('pegawai.update');
-        Route::delete('/pegawai/{id}','PegawaiDelete')->name('pegawai.delete');
+        Route::get('/admin/pegawai','PegawaiIndex')->name('pegawai');
+        Route::get('/admin/pegawai/tambah','PegawaitampilTambah')->name('pegawai.tambah');
+        Route::post('/admin/pegawai/tambah','PegawaiStore')->name('pegawai.post');
+        Route::get('/admin/pegawai/edit/{id}', 'PegawaiEdit')->name('pegawai.edit');
+        Route::put('/admin/pegawai/edit/{id}', 'PegawaiUpdate')->name('pegawai.update');
+        Route::delete('/admin/pegawai/{id}','PegawaiDelete')->name('pegawai.delete');
 
         Route::get('/export-pegawai/view/pdf','viewpdfPegawai')->name('export-pegawai');
 
         // JABATAN ROUTE
-        Route::get('/jabatan','JabatanIndex')->name('jabatan');
-        Route::post('/jabatan','JabatanStore')->name('jabatan.post');
-        Route::delete('/jabatan/{id}','JabatanDelete')->name('jabatan.delete');
-        Route::get('/jabatan/edit/{id}','JabatanEdit')->name('jabatan.edit');
-        Route::put('/jabatan/edit/{id}','JabatanUpdate')->name('jabatan.update');
+        Route::get('/admin/jabatan','JabatanIndex')->name('jabatan');
+        Route::post('/admin/jabatan','JabatanStore')->name('jabatan.post');
+        Route::delete('/admin/jabatan/{id}','JabatanDelete')->name('jabatan.delete');
+        Route::get('/admin/jabatan/edit/{id}','JabatanEdit')->name('jabatan.edit');
+        Route::put('/admin/jabatan/edit/{id}','JabatanUpdate')->name('jabatan.update');
 
         // BAGIAN ROUTE
-        Route::get('/bagian','BagianIndex')->name('bagian');
-        Route::post('/bagian','BagianStore')->name('bagian.post');
-        Route::delete('/bagian/{id}','BagianDelete')->name('bagian.delete');
-        Route::get('/bagian/edit/{id}','BagianEdit')->name('bagian.edit');
-        Route::put('/bagian/edit/{id}','BagianUpdate')->name('bagian.update');
+        Route::get('/admin/bagian','BagianIndex')->name('bagian');
+        Route::post('/admin/bagian','BagianStore')->name('bagian.post');
+        Route::delete('/admin/bagian/{id}','BagianDelete')->name('bagian.delete');
+        Route::get('/admin/bagian/edit/{id}','BagianEdit')->name('bagian.edit');
+        Route::put('/admin/bagian/edit/{id}','BagianUpdate')->name('bagian.update');
 
         // SHIFT ROUTE
-        Route::get('/shift','ShiftIndex')->name('shift');
-        Route::post('/shift','ShiftStore')->name('shift.post');
-        Route::delete('/shift/{id}','ShiftDelete')->name('shift.delete');
-        Route::get('/shift/edit/{id}','ShiftEdit')->name('shift.edit');
-        Route::put('/shift/edit/{id}','ShiftUpdate')->name('shift.update');
+        Route::get('/admin/shift','ShiftIndex')->name('shift');
+        Route::post('/admin/shift','ShiftStore')->name('shift.post');
+        Route::delete('/admin/shift/{id}','ShiftDelete')->name('shift.delete');
+        Route::get('/admin/shift/edit/{id}','ShiftEdit')->name('shift.edit');
+        Route::put('/admin/shift/edit/{id}','ShiftUpdate')->name('shift.update');
 
     });
 
     Route::controller(AttendancesController::class)->group(function () {
 
-        Route::get('/attendances/list','AttendanceInList')->name('listAttendances');
-        Route::get('/attendances','AttendanceIn')->name('attendances.in');
-        Route::post('/attendances','AttendanceInStore')->name('attendances.inPost');
-        Route::get('/attendances/out','AttendanceOut')->name('attendances.out');
-        Route::post('/attendances/out','AttendanceOutStore')->name('attendances.outPost');
-        Route::get('/attendances/detail/week','AttendancesDetail')->name('detailAttendances');
-        Route::get('/attendances/detail/month','AttendancesDetailMonth')->name('detailAttendancesMonth');
+        Route::get('/admin/attendances/list','AttendanceInList')->name('listAttendances');
+        Route::get('/admin/attendances','AttendanceIn')->name('attendances.in');
+        Route::post('/admin/attendances','AttendanceInStore')->name('attendances.inPost');
+        Route::get('/admin/attendances/out','AttendanceOut')->name('attendances.out');
+        Route::post('/admin/attendances/out','AttendanceOutStore')->name('attendances.outPost');
+        Route::get('/admin/attendances/detail/week','AttendancesDetail')->name('detailAttendances');
+        Route::get('/admin/attendances/detail/month','AttendancesDetailMonth')->name('detailAttendancesMonth');
 
         Route::get('/export-absensi-minggu/view/pdf','viewpdfAbsenWeek')->name('export-absensi-minggu');
         Route::get('/export-absensi-bulan/view/pdf','viewpdfAbsenMonth')->name('export-absensi-bulan');
@@ -84,8 +93,8 @@ Route::middleware(['auth'])->group(function () {
     });
 
 
-    Route::get('/rekapdata', function () {
-        return view('layout.rekapData');
+    Route::get('/admin/rekapdata', function () {
+        return view('admin.rekapData');
     })->name('rekapdata');
 
     // Route::get('/tugas', function () {
@@ -93,15 +102,15 @@ Route::middleware(['auth'])->group(function () {
     // });
 
 
-    Route::resource('manageuser', UserController::class);
-    Route::resource('sop', SopController::class);
-    Route::get('/sop/detail/{id}', [SopController::class, 'detailSop'])->name('sop.detail');
-    Route::get('/sop/pdf/{id}', [SopController::class, 'viewPdfSop'])->name('sop.pdf');
+    Route::resource('/admin/manageuser', UserController::class);
+    Route::resource('/admin/sop', SopController::class);
+    Route::get('/admin/sop/detail/{id}', [SopController::class, 'detailSop'])->name('sop.detail');
+    Route::get('/admin/sop/pdf/{id}', [SopController::class, 'viewPdfSop'])->name('sop.pdf');
 
-    Route::resource('task', TaskController::class);
-    Route::post('/task/create', [TaskController::class, 'storePegawai'])->name('task.pegawai');
+    Route::resource('/admin/task', TaskController::class);
+    Route::post('/admin/task/create', [TaskController::class, 'storePegawai'])->name('task.pegawai');
 
-    Route::resource('barcode', BarcodeController::class);
+    Route::resource('/admin/barcode', BarcodeController::class);
     Route::get('/download-qrcode', [BarcodeController::class ,'downloadQr'])->name('barcode.download');
     Route::get('/downloadAll-qrcode', [BarcodeController::class ,'downloadAll'])->name('barcode.downloadAll');
